qmmf-sdk: Enhancements to qmmf-sdk.
This commit has following changes.

- AVCodec adaptor layer to support video encode.
- AVCodec integration with Recorder, end to end implementation
  of record pipeline.
- AudioClient and AudioService.
- Code refactor, and clean up.
- Logging enhancements.
- Cmd line sample app enhancement to demonstrate usage of sdk Apis.
- Gtest framework enhancements.
- Standalone test apps to test AVCodec and AudioService functionality.
- Non-zsl snapshot support.

+++ recorder/src/service/qmmf_remote_cb.cpp
@@ -29,8 +29,8 @@
 
 #define TAG "RecorderRemoteCallBack"
 
-#include "qmmf_recorder_common.h"
-#include "qmmf_remote_cb.h"
+#include "recorder/src/service/qmmf_recorder_common.h"
+#include "recorder/src/service/qmmf_remote_cb.h"
 
 namespace qmmf {
 
@@ -66,7 +66,7 @@ void RemoteCallBack::NotifySessionEvent(EventType event_type, void *event_data,
                                         event_data_size);
 }
 
-void RemoteCallBack::NotifySnapshotData(void *buffer, size_t buffer_size) {
+void RemoteCallBack::NotifySnapshotData(void *buffer, uint32_t buffer_size) {
 
   assert(client_cb_handle_.get() != NULL);
   client_cb_handle_->NotifySnapshotData(buffer, buffer_size);
